+++ Auth-Guard/src/main/java/com/goormgb/be/authguard/jwt/repository/AccessTokenBlacklistRepository.java
@@ -1,4 +1,4 @@
-package com.goormgb.be.global.jwt.repository;
+package com.goormgb.be.authguard.jwt.repository;
 
 import java.time.Duration;
