+++ common-core/build.gradle
@@ -12,11 +12,6 @@ dependencies {
 	api 'org.springframework.boot:spring-boot-starter-webmvc'
 	api 'org.springdoc:springdoc-openapi-starter-webmvc-ui:3.0.1'
 
-	// JWT
-	api 'io.jsonwebtoken:jjwt-api:0.13.0'
-	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.13.0'
-	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.13.0'
-
 	runtimeOnly 'org.postgresql:postgresql'
 
 	// 필수: Actuator & Prometheus
